pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = [
    "hatchling",
]

[project]
name = "aa_bb"
description = "BigBrother"
readme = "README.md"
keywords = [
    "allianceauth",
    "aa_bb",
]
license = { file = "LICENSE" }
authors = [
    { name = "BioBrute", email = "broodlk@gmail.com"},
    { name = "Andrew Xadi", email = "andrewxadi@gmail.com" },
]
requires-python = ">=3.10"
classifiers = [
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]
dynamic = [
    "version",
]
dependencies = [
    "psycopg2-binary>=2.9.9",
    "matplotlib>=3.7.2",
    "allianceauth>=4.3.1",
    "allianceauth-corptools>=2.12.0",
    "allianceauth-discordbot >= 4.1.0",
    "aa-charlink >= 1.11.1",
    "django-esi >= 8.2.0"
]
urls.Changelog = "https://github.com/BroodLK/aa_bb/blob/main/CHANGELOG.md"
urls."Issue / Bug Reports" = "https://github.com/BroodLK/aa_bb/issues"

[tool.hatch.version]
path = "aa_bb/__init__.py"

[tool.hatch.build]
include = [
    "aa_bb/**"
]
